Move displayBoard function to module
