+++ .eslintrc.json
@@ -19,7 +19,7 @@
     }]
   },
   "ignorePatterns": [
-    "dist/",
+    "docs/",
     "build/"
   ]
 }
