+++ .stylelintrc.json
@@ -6,5 +6,5 @@
     "scss/at-rule-no-unknown": true,
     "csstree/validator": true
   },
-  "ignoreFiles": ["build/**", "docs/**", "**/reset*.css", "**/bootstrap*.css", "**/*.js", "**/*.jsx"]
+  "ignoreFiles": ["build/**", "docs/**", "**/reset*.css", "**/bootstrap*.css", "**/*.js", "**/*.jsx", "src/scss/style.css"]
 }
